requirements.txt
freezegun-1.0.0
mypy==0.782
pycodestyle==2.6.0
pylint==2.6.0
pytest-benchmark==3.2.3
pytest-cov==2.10.1
pytest-runner==5.2
pytest==6.0.1
radon==4.3.2
